# doc/api.yaml
openapi: 3.0.2
info:
  description: API for WasaPhoto project
  version: 1.0.0
  title: WasaPhoto API
  contact:
    email: matiddi.1985899@studenti.uniroma1.it
servers:
  - url: http://localhost:3000
...
paths:
  /session/:
    post:
      security: []
      tags: ["login", "users"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      requestBody:
        description: |-
          User details for login, composed
          only by the string representing the username.
        content:
          application/json:
            schema:
              type: object
              description: Object containing username string.
              properties:
                username:
                  type: string
                  example: "Maria"
                  pattern: ".*"
                  description: Username of the user trying to login.
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        "201":
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                description: Object containing user identifier.
                properties:
                  identifier:
                    description: |-
                      User unique identifier, 
                      integer utilization that corresponds 
                      as the one on the path.
                    type: integer
                    example: 1
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /photos/{photoId}/comments/:
    parameters:
      - $ref: "#/components/parameters/photoId"

    get:
      operationId: getComments
      description: |- 
        Get all the comments of the 
        photo to show them when opening the modal.
      summary: Get comments
      tags:
        - comments
      responses:
        "200":
          description: Comment list returned successfully
          content:
            application/json:
              schema:
                type: object
                description: Object containing the comment list.
                properties:
                  CList:
                    type: array
                    description: The comments of the photo
                    items: { $ref: "#/components/schemas/CommentListElement" }
                    minItems: 0
                    maxItems: 50
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      operationId: commentPhoto
      description: Write a comment to add to the photo selected.
      summary: Add comment
      tags:
        - comments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comment"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /users/{userId}:
    parameters:
      - $ref: "#/components/parameters/userId"

    put:
      description: Set a new username for the user you are logged in as.
      summary: Set new username
      operationId: setMyUserName
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Object containing new username
              properties:
                username:
                  type: string
                  description: New name for the user.
                  minLength: 0
                  maxLength: 100
                  example: gabrimat
                  pattern: ".*"
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      operationId: getUserProfile
      tags:
        - users
      description: |-
        Obtain photos published by the user selected via
        the parameter in the link and also other data to show.
      summary: Get the user profile
      responses:
        "200":
          description: User profile returned successfully.
          content:
            application/json:
              schema:
                type: object
                description: The object containing all necessary elements to show the user profile.
                properties:
                  PList:
                    type: array
                    description: The ids of the photos in the array.
                    items: { $ref: "#/components/schemas/PhotoListElement" }
                    minItems: 0
                    maxItems: 20
                  userName:
                    type: string
                    description: Name of the user.
                    example: Maria
                    pattern: ".*"
                    minLength: 3
                    maxLength: 16
                  followCount:
                    type: integer
                    example: 0
                    description: Number of users that follow the user.
                  followedCount:
                    type: integer
                    example: 0
                    description: Number of users that the user follows.
                  photoCount:
                    type: integer
                    example: 0
                    description: Number of photos posted by the user.
                  isBanned:
                    type: boolean
                    example: true
                    description: If you banned the user.
                  isFollowed:
                    type: boolean
                    example: false
                    description: If you follow the user.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/stream/:
    parameters:
      - $ref: "#/components/parameters/userId"

    get:
      summary: Get user's feed
      description: Retrieve the feed of the specified user.
      operationId: getMyStream
      tags:
        - users
      responses:
        "200":
          description: Photo list returned successfully.
          content:
            application/json:
              schema:
                type: object
                description: Object containing elements of the feed.
                properties:
                  PList:
                    type: array
                    description: Array contain information to show for each photo.
                    items: { $ref: "#/components/schemas/PhotoStreamListElement" }
                    minItems: 0
                    maxItems: 20
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/follows/{followId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/followId"

    put:
      description: |-
        Put a follow with the id of the user
        followed in the follows of the user.
      summary: Follow a user
      operationId: followUser
      tags:
        - follows
      responses:
        "200":
          $ref: "#/components/responses/Created"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      description: Stop following the selected user.
      summary: Unfollow a user
      operationId: unfollowUser
      tags:
        - follows
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/bans/{banId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/banId"

    put:
      description: |-
        Ban the user you are currently viewing,
        stop showing your content and profile to him.
      summary: Ban a user
      operationId: banUser
      tags:
        - bans
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      description: Stop the ban relation with the user you are currently viewing.
      summary: Unban a user
      operationId: unbanUser
      tags:
        - bans
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  responses:
    OK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResponseMessage"
    Created:
      description: Created response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ResponseMessage"
    NotFound:
      description: Not Found response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "NOT_FOUND"
              message: "Element not found."
    Unauthorized:
      description: Access token missing or not valid response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Authorization token invalid or not present."
    Forbidden:
      description: Forbidden response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "FORBIDDEN"
              message: "Wrong authorization for the operation."
...
    InternalServerError:
      description: Internal server error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "INTERNAL_SERVER_ERROR"
              message: "An internal server error occurred."
  parameters:
...
    photoId:
      in: path
      name: photoId
      required: true
      description: Photo identifier in the path.
      schema:
        description: Integer for the identifier of the photo.
        type: integer
    userId:
      in: path
      name: userId
      required: true
      description: Identifier of the user in the path.
      schema:
        description: Integer representing the identifier of the user.
        type: integer
    followId:
      in: path
      name: followId
      required: true
      description: User identifier for follower in the path.
      schema:
        description: Id of the user at the end of the follow relation.
        type: integer
    banId:
      in: path
      name: banId
      required: true
      description: User identifier for banned user.
      schema:
        description: Id of the user at the end of the ban relation.
        type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: integer
  schemas:
    Error:
      type: object
      description: Error response made of error code and message.
      properties:
        code:
          type: string
          description: Code of the error.
          minLength: 0
          maxLength: 4
          pattern: ".*"
        message:
          type: string
          description: Message of the error.
          minLength: 0
          maxLength: 50
          pattern: ".*"
    ResponseMessage:
      type: object
      description: |-
        Response
        after endpoint running.
...
    PhotoListElement:
      title: PhotoListElement
      description: |-
        This object is a component of the photo list
        returned to represent a photo in the user profile.
      type: object
      properties:
        id:
          type: integer
          example: 1
          description: Identifier of the photo.
        likeCount:
          type: integer
          example: 0
          description: Amount of likes of the photo.
        publisherId:
          type: integer
          example: 2
          description: Identifier of the user that published the photo.
        caption:
          type: string
          description: Caption of the photo.
          example: Good day today!
          maxLength: 1000
          minLength: 0
          pattern: ".*"
        releaseDate:
          type: string
          format: date-time
          example: "2017-07-21T17:32:28Z"
          minLength: 20
          maxLength: 20
          description: Date of the photo upload.
    PhotoStreamListElement:
      title: PhotoListElement
      description: |-
        This object is a component of the photo
        list returned when viewing you stream.
      type: object
      properties:
        id:
          type: integer
          example: 1
          description: Identifier of the photo.
        likeCount:
          type: integer
          example: 0
          description: Amount of likes.
        publisherName:
          type: string
          example: gabrimat
          description: Name of the user that published the photo.
          minLength: 3
          maxLength: 16
          pattern: ".*"
        caption:
          type: string
          description: Caption of the photo in the stream.
          example: good day!
          maxLength: 1000
          minLength: 0
          pattern: ".*"
        releaseDate:
          type: string
          format: date-time
          example: "2017-07-21T17:32:28Z"
          minLength: 20
          maxLength: 20
          description: Date of the photo release.
    Comment:
      title: Comment
      description: Comment of a published photo.
      type: object
      properties:
        commentText:
          description: Text content of the comment.
          type: string
          example: Nice photo!
          minLength: 0
          maxLength: 1000
          pattern: ".*"
    CommentListElement:
      description: Element of the comment list.
      title: CommentListElement
      type: object
      properties:
        id:
          description: Identifier of the comment.
          type: integer
          example: 1
        photoId:
          description: Identifier of the photo which the comment belongs to.
          example: 1
          type: integer
        publisherId:
          description: Identifier of the user that published the comment.
          example : 1
          type: integer
        publisherName:
          description: Name of the user that published the comment.
          type: string
          example: Maria
          minLength: 3
          maxLength: 16
          pattern: ".*"
        commentText:
          description: Text content of the comment.
          type: string
          example: Nice photo!
          minLength: 0
          maxLength: 1000
          pattern: ".*"
security:
  - bearerAuth: []